🐛 Fix: modal layout bug

+++ src/components/dashboardEdit/Dashboard.tsx
@@ -6,6 +6,7 @@ import { useRouter } from 'next/router';
 import { useForm } from 'react-hook-form';
 import { handleCancelInvitation, handleDeleteMember } from '@/src/pages/api/dashboardEditApi';
 import { useDashboardStore, useMembersStore, useInviteesStore } from '@/src/util/zustand';
+import useModal from '@/src/hooks/useModal';
 import ColorPicker from '../common/colorpicker';
 import Button from '../common/button';
 import Table from './table';
@@ -16,6 +17,7 @@ import Input, { InputForm } from '../common/input';
 import MembersPagination from './table/MembersPagination';
 import InviteePagination from './table/InviteePagination';
 import InviteModal from '../InviteModal';
+import ModalPortal from '../common/modalPortal';
 
 const Dashboard = () => {
   const dashboard = useDashboardStore((state) => state.dashboardData);
@@ -27,13 +29,7 @@ const Dashboard = () => {
   const router = useRouter();
   const { id } = router.query;
   const idNumber = Number(id);
-  const [isAddDashboardModalVisible, setIsAddDashboardModalVisible] = useState(false);
-  const showAddDashboardModal = () => {
-    setIsAddDashboardModalVisible(true);
-  };
-  const hideAddDashboardModal = () => {
-    setIsAddDashboardModalVisible(false);
-  };
+  const { openModal: inviteModal, handleModalClose: inviteModalClose, handleModalOpen: inviteModalOpen } = useModal();
   const removeInvitee = useInviteesStore((state) => state.removeInvitee);
   const removeMember = useMembersStore((state) => state.removeMember);
   const handleEditDashboard = async () => {
@@ -47,7 +43,11 @@ const Dashboard = () => {
   };
 
   const handleDeleteDashboard = async () => {
+    const confirmDeletion = window.confirm('대시보드를 삭제할까요?');
+    if (!confirmDeletion) return;
+
     await instance.delete(`/dashboards/${idNumber}`);
+    alert('대시보드를 삭제했습니다.');
     router.push('/my-dashboard');
   };
 
@@ -118,7 +118,7 @@ const Dashboard = () => {
             bgColor="violet"
             textColor="white"
             className="mobile:absolute mobile:right-20 mobile:top-90 w-105"
-            onClick={showAddDashboardModal}
+            onClick={inviteModalOpen}
           >
             <div className="flex gap-6 items-center justify-center ">
               <Image src={add} alt="add" />
@@ -149,9 +149,9 @@ const Dashboard = () => {
       <Button buttonType="dashboardDelete" bgColor="white" className="mt-25" onClick={() => handleDeleteDashboard()}>
         대시보드 삭제하기
       </Button>
-      {isAddDashboardModalVisible && (
-        <InviteModal openModal={isAddDashboardModalVisible} handleModalClose={hideAddDashboardModal} />
-      )}
+      <ModalPortal>
+        <InviteModal openModal={inviteModal} handleModalClose={inviteModalClose} />
+      </ModalPortal>
     </div>
   );
 };
